신규: #16 선수 기초 데이터 동기화 기능 구현

+++ league-service/src/main/java/com/kickoff/service/league/domain/service/usecase/team/InitTeamsUseCase.java
@@ -17,7 +17,7 @@ public class InitTeamsUseCase {
   private final StreamBridge streamBridge;
 
   public void fetchSeasons(InitTeamsCommand initTeamsCommand) {
-    League league = leagueRepository.findByApiFootballLeagueId((long) initTeamsCommand.getApiFootballLeagueId())
+    League league = leagueRepository.findByApiFootballLeagueId(initTeamsCommand.getApiFootballLeagueId())
       .orElseThrow();
     initTeamsCommand.setLeagueId(league.getId());
     league.getSeasons().forEach(season -> initTeamsCommand.addSeason(season.getId().getYear()));
@@ -25,7 +25,7 @@ public class InitTeamsUseCase {
   }
 
   public void persistSeasonMapTeamsForInitTeams(InitTeamsCommand initTeamsCommand) {
-    League league = leagueRepository.findByApiFootballLeagueId((long) initTeamsCommand.getApiFootballLeagueId()).orElseThrow();
+    League league = leagueRepository.findByApiFootballLeagueId(initTeamsCommand.getApiFootballLeagueId()).orElseThrow();
     for (Year year : initTeamsCommand.getYears()) {
       initTeamsCommand.getTeams(year).forEach(teamId -> league.addTeam(year, teamId));
     }
